Remove import/export monitoring
These bots are no longer the responsibility of the Interop Tooling team.
Export PRs getting stuck is still our responsibility, however.

That that what remains could now be a static web app, but leave
everything as-is, as what remains will likely be folded into some other
rotation anyway and this tool discontinued.