Find import/export jobs via Buildbucket RPC (#74)
The builder page was recently redesigned, which broke the `wptsync`
scraper. This causes `/api/wptsync/*` to never return any builds, so the
tasks on the dashboard are identically green.

This PR updates the API handler to fetch import/export job data via
Buildbucket RPC, which is designed for machine consumption and therefore
more robust. There should be no functional change to `/apt/wptsync/*`.

This also obviates the need for MechanicalSoup for scraping, so that
dependency has been removed.