Update ecosystem-infra-rotation.html
Updates the blocked export PR query to only include the PRs that are only approved.

In particular, this updated query will exclude PRs that still need manual review by the Interop team.

Question for the reviewer: Do we still want to track those PRs where human review is still needed? Maybe because of some reason they need a nudge from us? If so, we should probably think of another way to handle this.